Update the NET API SDK form OPCF - entered 2.01.105.0 from NET API 2.00 Source Code 105.1 dated: ‎‎2.01.105.1
1. Upgrades the projects to VS2008
2. Uses to the platform independent RCW included in Core Components Build 105.0
3. Fixed outstanding mantis issues.

The new RCW changed some APIs so the behavior of applications could change.
The assembly version has been changed to 2.01.105.0. Applications will need to be re-built to use this version.
Finalizers were added to classes that implement IDisposeable. This could cause a change in application behavior.
The Merge Module installs the development environment if the OPC_INSTALL_SDK property is set.

Issue #: 4842